StepHistoryList -> StepStack
