--- javascript/react/docs/markdown/react.stephistorylist.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryList](./react.stephistorylist.md)
-
-## StepHistoryList variable
-
-<b>Signature:</b>
-
-```typescript
-StepHistoryList: React.FC<StepHistoryListProps>
-```
