--- javascript/react/docs/markdown/react.stephistorylistprops.md
@@ -1,25 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md)
-
-## StepHistoryListProps interface
-
-<b>Signature:</b>
-
-```typescript
-export interface StepHistoryListProps 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [onDelete?](./react.stephistorylistprops.ondelete.md) |  | (index: number) =&gt; void | <i>(Optional)</i> |
-|  [onSave?](./react.stephistorylistprops.onsave.md) |  | (step: Step, index?: number) =&gt; void | <i>(Optional)</i> |
-|  [onSelect?](./react.stephistorylistprops.onselect.md) |  | (name: string) =&gt; void | <i>(Optional)</i> |
-|  [order?](./react.stephistorylistprops.order.md) |  | [DisplayOrder](./react.displayorder.md) | <i>(Optional)</i> |
-|  [selectedKey?](./react.stephistorylistprops.selectedkey.md) |  | string | <i>(Optional)</i> Id of the selected step |
-|  [showSelectButtons?](./react.stephistorylistprops.showselectbuttons.md) |  | boolean | <i>(Optional)</i> Indicates whether the original/latest buttons will be shown. Note that they will have no effect if you do not supply an onSelect handler too. |
-|  [styles?](./react.stephistorylistprops.styles.md) |  | StepHistoryStyles | <i>(Optional)</i> |
-|  [workflow](./react.stephistorylistprops.workflow.md) |  | Workflow |  |
-
