--- javascript/react/docs/markdown/react.stephistorylistprops.ondelete.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [onDelete](./react.stephistorylistprops.ondelete.md)
-
-## StepHistoryListProps.onDelete property
-
-<b>Signature:</b>
-
-```typescript
-onDelete?: (index: number) => void;
-```
