--- javascript/react/docs/markdown/react.stephistorylistprops.onsave.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [onSave](./react.stephistorylistprops.onsave.md)
-
-## StepHistoryListProps.onSave property
-
-<b>Signature:</b>
-
-```typescript
-onSave?: (step: Step, index?: number) => void;
-```
