--- javascript/react/docs/markdown/react.stephistorylistprops.onselect.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [onSelect](./react.stephistorylistprops.onselect.md)
-
-## StepHistoryListProps.onSelect property
-
-<b>Signature:</b>
-
-```typescript
-onSelect?: (name: string) => void;
-```
