--- javascript/react/docs/markdown/react.stephistorylistprops.order.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [order](./react.stephistorylistprops.order.md)
-
-## StepHistoryListProps.order property
-
-<b>Signature:</b>
-
-```typescript
-order?: DisplayOrder;
-```
