--- javascript/react/docs/markdown/react.stephistorylistprops.selectedkey.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [selectedKey](./react.stephistorylistprops.selectedkey.md)
-
-## StepHistoryListProps.selectedKey property
-
-Id of the selected step
-
-<b>Signature:</b>
-
-```typescript
-selectedKey?: string;
-```
