--- javascript/react/docs/markdown/react.stephistorylistprops.styles.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [styles](./react.stephistorylistprops.styles.md)
-
-## StepHistoryListProps.styles property
-
-<b>Signature:</b>
-
-```typescript
-styles?: StepHistoryStyles;
-```
