--- javascript/react/docs/markdown/react.stephistorylistprops.workflow.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [workflow](./react.stephistorylistprops.workflow.md)
-
-## StepHistoryListProps.workflow property
-
-<b>Signature:</b>
-
-```typescript
-workflow: Workflow;
-```
