+++ javascript/react/docs/markdown/react.stepstack.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStack](./react.stepstack.md)
+
+## StepStack variable
+
+<b>Signature:</b>
+
+```typescript
+StepStack: React.FC<StepStackProps>
+```
