+++ javascript/react/docs/markdown/react.stepstackprops.md
@@ -1,0 +1,25 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md)
+
+## StepStackProps interface
+
+<b>Signature:</b>
+
+```typescript
+export interface StepStackProps 
+```
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [onDelete?](./react.stepstackprops.ondelete.md) |  | (index: number) =&gt; void | <i>(Optional)</i> |
+|  [onSave?](./react.stepstackprops.onsave.md) |  | (step: Step, index?: number) =&gt; void | <i>(Optional)</i> |
+|  [onSelect?](./react.stepstackprops.onselect.md) |  | (name: string) =&gt; void | <i>(Optional)</i> |
+|  [order?](./react.stepstackprops.order.md) |  | [DisplayOrder](./react.displayorder.md) | <i>(Optional)</i> |
+|  [selectedKey?](./react.stepstackprops.selectedkey.md) |  | string | <i>(Optional)</i> Id of the selected step |
+|  [showSelectButtons?](./react.stepstackprops.showselectbuttons.md) |  | boolean | <i>(Optional)</i> Indicates whether the original/latest buttons will be shown. Note that they will have no effect if you do not supply an onSelect handler too. |
+|  [styles?](./react.stepstackprops.styles.md) |  | StepHistoryStyles | <i>(Optional)</i> |
+|  [workflow](./react.stepstackprops.workflow.md) |  | Workflow |  |
+
