+++ javascript/react/docs/markdown/react.stepstackprops.ondelete.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [onDelete](./react.stepstackprops.ondelete.md)
+
+## StepStackProps.onDelete property
+
+<b>Signature:</b>
+
+```typescript
+onDelete?: (index: number) => void;
+```
