+++ javascript/react/docs/markdown/react.stepstackprops.onsave.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [onSave](./react.stepstackprops.onsave.md)
+
+## StepStackProps.onSave property
+
+<b>Signature:</b>
+
+```typescript
+onSave?: (step: Step, index?: number) => void;
+```
