+++ javascript/react/docs/markdown/react.stepstackprops.onselect.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [onSelect](./react.stepstackprops.onselect.md)
+
+## StepStackProps.onSelect property
+
+<b>Signature:</b>
+
+```typescript
+onSelect?: (name: string) => void;
+```
