+++ javascript/react/docs/markdown/react.stepstackprops.order.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [order](./react.stepstackprops.order.md)
+
+## StepStackProps.order property
+
+<b>Signature:</b>
+
+```typescript
+order?: DisplayOrder;
+```
