+++ javascript/react/docs/markdown/react.stepstackprops.selectedkey.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [selectedKey](./react.stepstackprops.selectedkey.md)
+
+## StepStackProps.selectedKey property
+
+Id of the selected step
+
+<b>Signature:</b>
+
+```typescript
+selectedKey?: string;
+```
