+++ javascript/react/docs/markdown/react.stepstackprops.showselectbuttons.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepHistoryListProps](./react.stephistorylistprops.md) &gt; [showSelectButtons](./react.stephistorylistprops.showselectbuttons.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [showSelectButtons](./react.stepstackprops.showselectbuttons.md)
 
-## StepHistoryListProps.showSelectButtons property
+## StepStackProps.showSelectButtons property
 
 Indicates whether the original/latest buttons will be shown. Note that they will have no effect if you do not supply an onSelect handler too.
