+++ javascript/react/docs/markdown/react.stepstackprops.styles.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [styles](./react.stepstackprops.styles.md)
+
+## StepStackProps.styles property
+
+<b>Signature:</b>
+
+```typescript
+styles?: StepHistoryStyles;
+```
