+++ javascript/react/docs/markdown/react.stepstackprops.workflow.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepStackProps](./react.stepstackprops.md) &gt; [workflow](./react.stepstackprops.workflow.md)
+
+## StepStackProps.workflow property
+
+<b>Signature:</b>
+
+```typescript
+workflow: Workflow;
+```
