--- javascript/react/src/components/StepHistoryList/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export { StepHistoryList } from './StepHistoryList.js'
-export type { StepHistoryListProps } from './StepHistoryList.types.js'
