+++ javascript/react/src/components/StepStack/StepStack.hooks.ts
