+++ javascript/react/src/components/StepStack/StepStack.styles.ts
