+++ javascript/react/src/components/StepStack/StepStack.utils.ts
