+++ javascript/react/src/components/StepStack/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export { StepStack } from './StepStack.js'
+export type { StepStackProps } from './StepStack.types.js'
