+++ javascript/react/src/components/index.ts
@@ -7,7 +7,7 @@ export * from './ArqueroTableHeader/index.js'
 export * from './Guidance/index.js'
 export * from './StepDescription/index.js'
 export * from './StepForm/index.js'
-export * from './StepHistoryList/index.js'
 export * from './StepSelector/index.js'
+export * from './StepStack/index.js'
 export * from './TableCommands/index.js'
 export * from './TableTransform/index.js'
